replace evaluate_ppo with dup_eval
